feat: parse certificate in server-side instead of client-side
