+++ crate/server/src/core/operations/import.rs
@@ -1,16 +1,25 @@
 use std::collections::HashSet;
 
 use cosmian_kmip::kmip::{
-    kmip_data_structures::KeyValue,
+    kmip_data_structures::{KeyBlock, KeyMaterial, KeyValue},
     kmip_objects::{Object, ObjectType},
     kmip_operations::{Import, ImportResponse},
-    kmip_types::{KeyWrapType, StateEnumeration},
+    kmip_types::{
+        Attributes, CertificateType, CryptographicAlgorithm, CryptographicDomainParameters,
+        CryptographicUsageMask, KeyFormatType, KeyWrapType, RecommendedCurve, StateEnumeration,
+    },
 };
 use cosmian_kms_utils::{
     access::ExtraDatabaseParams,
+    crypto::curve_25519::operation::Q_LENGTH_BITS,
     tagging::{check_user_tags, get_tags},
 };
-use tracing::{debug, warn};
+use openssl::{
+    ec::EcKey,
+    nid::Nid,
+    pkey::{Id, PKey},
+};
+use tracing::debug;
 use x509_parser::{parse_x509_certificate, prelude::parse_x509_pem};
 
 use super::wrapping::unwrap_key;
@@ -57,6 +66,96 @@ fn parse_certificate_and_create_tags(
     Ok(())
 }
 
+fn get_private_key_object(
+    private_key_bytes: Vec<u8>,
+    recommended_curve: RecommendedCurve,
+) -> Object {
+    Object::PrivateKey {
+        key_block: KeyBlock {
+            key_format_type: KeyFormatType::TransparentECPrivateKey,
+            key_value: KeyValue {
+                key_material: KeyMaterial::ByteString(private_key_bytes),
+                attributes: Some(Attributes {
+                    activation_date: None,
+                    cryptographic_algorithm: Some(CryptographicAlgorithm::ECDH),
+                    cryptographic_length: Some(Q_LENGTH_BITS),
+                    cryptographic_domain_parameters: Some(CryptographicDomainParameters {
+                        q_length: Some(Q_LENGTH_BITS),
+                        recommended_curve: Some(recommended_curve),
+                    }),
+                    cryptographic_parameters: None,
+                    cryptographic_usage_mask: Some(
+                        CryptographicUsageMask::Encrypt
+                            | CryptographicUsageMask::Decrypt
+                            | CryptographicUsageMask::WrapKey
+                            | CryptographicUsageMask::UnwrapKey
+                            | CryptographicUsageMask::KeyAgreement,
+                    ),
+                    key_format_type: Some(KeyFormatType::ECPrivateKey),
+                    link: None,
+                    object_type: Some(ObjectType::PrivateKey),
+                    vendor_attributes: None,
+                }),
+            },
+            cryptographic_algorithm: CryptographicAlgorithm::ECDH,
+            cryptographic_length: Q_LENGTH_BITS,
+            key_compression_type: None,
+            key_wrapping_data: None,
+        },
+    }
+}
+
+fn import_pem(tags: &mut HashSet<String>, pem_value: &[u8]) -> KResult<Object> {
+    let (_, pem) = parse_x509_pem(pem_value)?;
+
+    let object = if pem.label == "CERTIFICATE" {
+        debug!("CLI: parsing certificate: {}", pem.label);
+        parse_certificate_and_create_tags(tags, pem_value)?;
+        Object::Certificate {
+            certificate_type: CertificateType::X509,
+            certificate_value: pem_value.to_vec(),
+        }
+    } else if pem.label.contains("PRIVATE KEY") {
+        debug!("CLI: parsing private key: {}", pem.label);
+        let pkey = PKey::private_key_from_pem(pem_value)?;
+        match pkey.id() {
+            Id::EC => {
+                debug!("CLI: parsing private key with PKey: {:?}", pkey);
+                let private_key = EcKey::private_key_from_der(&pem.contents)?;
+                debug!("CLI: convert private key to EcKey");
+                let recommended_curve = match private_key.group().curve_name() {
+                    Some(nid) => match nid {
+                        Nid::X9_62_PRIME192V1 => RecommendedCurve::P192,
+                        Nid::SECP224R1 => RecommendedCurve::P224,
+                        Nid::X9_62_PRIME256V1 => RecommendedCurve::P256,
+                        Nid::SECP384R1 => RecommendedCurve::P384,
+                        _ => {
+                            kms_bail!("Elliptic curve not supported: {}", nid.long_name()?);
+                        }
+                    },
+                    None => kms_bail!("No curve name for this EC curve"),
+                };
+                let private_key_bytes = private_key.private_key().to_vec();
+                debug!("CLI: private_key_bytes len: {}", private_key_bytes.len());
+                get_private_key_object(private_key_bytes, recommended_curve)
+            }
+            Id::ED25519 => {
+                let private_key_bytes = pkey.raw_private_key()?;
+                get_private_key_object(private_key_bytes, RecommendedCurve::CURVEED25519)
+            }
+            Id::X25519 => {
+                let private_key_bytes = pkey.raw_private_key()?;
+                get_private_key_object(private_key_bytes, RecommendedCurve::CURVE25519)
+            }
+            _ => kms_bail!("Private key id not supported: {:?}", pkey.id()),
+        }
+    } else {
+        kms_bail!("Unsupported PEM format: found {}", pem.label);
+    };
+
+    Ok(object)
+}
+
 /// Import a new object
 pub async fn import(
     kms: &KMS,
@@ -77,10 +176,10 @@ pub async fn import(
     let mut tags = get_tags(&request.attributes);
     check_user_tags(&tags)?;
 
-    let mut object = request.object;
-    let object_type = object.object_type();
-    match object_type {
+    let object_type = request.object.object_type();
+    let mut object = match object_type {
         ObjectType::SymmetricKey | ObjectType::PublicKey | ObjectType::PrivateKey => {
+            let mut object = request.object;
             let object_key_block = object.key_block_mut()?;
             // unwrap before storing if requested
             if request.key_wrap_type == Some(KeyWrapType::NotWrapped) {
@@ -104,11 +203,12 @@ pub async fn import(
                 }
                 _ => unreachable!(),
             }
+            object
         }
         ObjectType::Certificate => {
             debug!("Import with _cert system tag");
             tags.insert("_cert".to_string());
-            let certificate_pem_bytes = match &object {
+            let certificate_pem_bytes = match &request.object {
                 Object::Certificate {
                     certificate_value, ..
                 } => Ok(certificate_value),
@@ -116,12 +216,14 @@ pub async fn import(
                     "Invalid object type {object_type:?} when importing a certificate"
                 ))),
             }?;
-            parse_certificate_and_create_tags(&mut tags, certificate_pem_bytes)?;
+            import_pem(&mut tags, certificate_pem_bytes)?
         }
         x => {
-            warn!("Import is not yet supported for objects of type : {x}");
+            return Err(KmsError::InvalidRequest(format!(
+                "Import is not yet supported for objects of type : {x}"
+            )))
         }
-    }
+    };
 
     // check if the object will be replaced if it already exists
     let replace_existing = if let Some(v) = request.replace_existing {
